refactor(website): split website editor utilities and shared components
